Make SaveProgressButton a modal
