+++ components/chapters/SignUpModal.tsx
@@ -1,0 +1,114 @@
+'use client'
+
+import Image from 'next/image'
+import { useState } from 'react'
+import Modal from 'react-modal'
+
+export const SignUpModal = (props) => {
+  function randomHex(length) {
+    let result = ''
+    let characters = '0123456789abcdef'
+    let charactersLength = characters.length
+    for (var i = 0; i < length; i++) {
+      result += characters.charAt(Math.floor(Math.random() * charactersLength))
+    }
+    return result
+  }
+
+  const [code, _] = useState(randomHex(64))
+
+  let [copied, setCopied] = useState(false)
+  let [avatar, setAvatar] = useState(1)
+
+  function copy() {
+    navigator.clipboard.writeText(code)
+    setCopied(true)
+
+    setTimeout(() => {
+      setCopied(false)
+    }, 2000)
+  }
+
+  return (
+    <Modal
+      isOpen={props.open}
+      className="top-1/2 left-1/2 outline-none p-5 transform -translate-x-1/2 -translate-y-1/2 absolute w-[32rem] rounded-lg shadow-lg bg-[#32455D] text-white"
+      contentLabel="Example Modal">
+
+        <div className="float-right flex justify-end">
+          <button onClick={props.onClose}>
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              className="h-6 w-6"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth={2}
+                d="M6 18L18 6M6 6l12 12"
+              />
+            </svg>
+          </button>
+        </div>
+
+      <h2 className="mb-4 text-xl font-bold">Want to save your progress?</h2>
+      <p className="mb-5 text-sm">
+        Copy and store a simple code to save and load your progress on any
+        device and browser. If you already have a code, load your progress
+        here.
+      </p>
+
+      <h2 className="mb-4 text-xl font-bold">Choose an avatar</h2>
+
+      <div className="mb-5 flex cursor-pointer justify-between">
+        {[1, 2, 3, 4, 5].map((i) => (
+        <div className="flex flex-col items-center" key={i}>
+          <Image
+            src={`/assets/avatars/${i}.png`}
+            alt="avatar"
+            width={80}
+            height={80}
+            onClick={() => setAvatar(i)}
+            className={`h-20 w-20 rounded-full ${avatar === i ? 'border-2 border-blue-500' : ''}`}
+          />
+        </div>
+        ))}
+      </div>
+
+      <h2 className="mb-4 text-xl font-bold">Back up your personal code</h2>
+
+      <pre className="mb-5 flex flex-col rounded-md border-2 border-dotted border-white/25 p-4">
+        <code className="whitespace-pre-wrap break-all">{code}</code>
+        {copied ? (
+          <button
+            className="mt-4 w-full rounded-md bg-green-500 px-4 py-2 text-green-800"
+            onClick={copy}
+          >
+            Copied!
+          </button>
+        ) : (
+          <button
+            className="mt-4 w-full rounded-md bg-[#233346] px-4 py-2 text-white"
+            onClick={copy}
+          >
+            Copy
+          </button>
+        )}
+      </pre>
+
+      <p className="mt-5 text-sm">
+        All set? Code copied and backed-up? Make sure your do, as it can’t be
+        recovered if you lose it.
+      </p>
+      <button
+        onClick={() => props.onConfirm({ code, avatar })}
+        className="mt-4 w-full rounded-md border border-white px-4 py-2 text-white"
+      >
+        I’ve copied and backed up my code
+      </button>
+    </Modal>
+  )
+}
